Review latest PR
GuiTextBox() redesign is a huge addition and extends old text box capabilities a lot... but it implied a big amount of new functions and code. Some user could not require that extended functionality, so, I allowed enabling it with a flag while keeping the old (and simpler) implementations.

I reviewed all the code formating but it requires a deeper analysis, probably some functions could be moved to raylib text module.