REVIEWED: Narrowing issue with styles #179
Actually this issue is a bit tricky, C++ (g++ only) complains about narrowing on styles if `propertyValue` is assigned an hexadecimal value because it was not `unsigned int`, making it unsigned it avoids that error but then properties can only be positive... maybe we need some property value to be negative?

Maybe this issue will require further review in the future.