Added GuiIntBox() and GuiFloatBox() functions
GuiIntBox() is basically the same thing as GuiValueBox(), but it works with negative values. Didn't want to delete GuiValueBox() to not cause comaptibility issues with projects that rely upon it. GuiFloatBox() is just a box for float values. It requires TextToFloat() to function properly, i've added implementation for it in GUI_STANDALONE section and will create commit to main raylib library with same function implementation in rtext.c file.